SRP in action; js_only.
