+++ assets/src/domain/shared/entities/datetimes/predicates/filters/utils.ts
@@ -1,0 +1,3 @@
+import { parseISO, formatISO } from 'date-fns';
+
+export const now = parseISO(formatISO(new Date()));
